fix: CHK moved in rawdata
