+++ backend/app/services/mapdata/io.py
@@ -8,7 +8,7 @@ from eudplib.maprw.loadmap import LoadMap
 from eudplib.bindings._rust import mpqapi
 from tempfile import NamedTemporaryFile
 from io import BytesIO
-from app.services.mapdata.chk import CHK, CHKBuilder
+from app.services.rawdata.chk import CHK, CHKBuilder
 from app.services.rawdata.dat import DAT
 from app.models.project import Usemap
 from app.services.bridge.transformer import Transformer
